third_party/blink/web_tests/virtual/webnn-service-on-cpu/external/wpt/webnn/conformance_tests/maxPool2d.https.any_cpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] maxPool2d float32 4D tensor options.dilations
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Pool2d in tflite doesn't support dilations."
[FAIL] [required] maxPool2d float32 4D tensor options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float32 actual 99.28312683105469 should be close enough to expected -39.03501892089844 by ULP distance: expected a number less than or equal to 0n but got 2229449938n
[FAIL] [required] maxPool2d float32 4D tensor options.outputSizes ignores options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float32 actual 99.28312683105469 should be close enough to expected -39.03501892089844 by ULP distance: expected a number less than or equal to 0n but got 2229449938n
[FAIL] [required] maxPool2d float16 4D tensor options.dilations
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Pool2d in tflite doesn't support dilations."
[FAIL] [required] maxPool2d float16 4D tensor options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 99.3125 should be close enough to expected -39.03125 by ULP distance: expected a number less than or equal to 0 but got 31404
[FAIL] [required] maxPool2d float16 4D tensor options.outputSizes ignores options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 99.3125 should be close enough to expected -39.03125 by ULP distance: expected a number less than or equal to 0 but got 31404
Harness: the test ran to completion.
